Add OpenSamlRelyingPartyRegistration
Issue gh-12841
